Fixed regression in Windows builds from Android updates

+++ iModelCore/WSClient/PublicAPI/WebServices/ConnectC/CWSCCGenBufferPublic.h
@@ -192,7 +192,7 @@ CWSCC_EXPORT CALLSTATUS ConnectWebServicesClientC_DataBufferGetStringProperty
 (
 CWSCCDATABUFHANDLE dataBuffer,
 int16_t bufferProperty,
-int16_t index,
+uint32_t index,
 WCharP str,
 uint32_t strLength
 );
@@ -209,7 +209,7 @@ CWSCC_EXPORT CALLSTATUS ConnectWebServicesClientC_DataBufferGetStringLength
 (
 CWSCCDATABUFHANDLE dataBuffer,
 int16_t bufferProperty,
-int16_t index,
+uint32_t index,
 size_t* outStringSize
 );
 
@@ -225,7 +225,7 @@ CWSCC_EXPORT CALLSTATUS ConnectWebServicesClientC_DataBufferGetGuidProperty
 (
 CWSCCDATABUFHANDLE dataBuffer,
 int16_t bufferProperty,
-int16_t index,
+uint32_t index,
 WCharP guid
 );
 
@@ -241,7 +241,7 @@ CWSCC_EXPORT CALLSTATUS ConnectWebServicesClientC_DataBufferGetBooleanProperty
 (
 CWSCCDATABUFHANDLE dataBuffer,
 int16_t bufferProperty,
-int16_t index,
+uint32_t index,
 bool* boolean
 );
 
@@ -257,8 +257,8 @@ CWSCC_EXPORT CALLSTATUS ConnectWebServicesClientC_DataBufferGetIntProperty
 (
 CWSCCDATABUFHANDLE dataBuffer,
 int16_t bufferProperty,
-int16_t index,
-int16_t* integer
+uint32_t index,
+int32_t* integer
 );
 
 /************************************************************************************//**
@@ -273,7 +273,7 @@ CWSCC_EXPORT CALLSTATUS ConnectWebServicesClientC_DataBufferGetDoubleProperty
 (
 CWSCCDATABUFHANDLE dataBuffer,
 int16_t bufferProperty,
-int16_t index,
+uint32_t index,
 double* pDouble
 );
 
@@ -289,8 +289,8 @@ CWSCC_EXPORT CALLSTATUS ConnectWebServicesClientC_DataBufferGetLongProperty
 (
 CWSCCDATABUFHANDLE dataBuffer,
 int16_t bufferProperty,
-int16_t index,
-int32_t* pLong
+uint32_t index,
+int64_t* pLong
 );
 
 /** \} */
